backend/requirements.txt
amqp==5.0.2
appdirs==1.4.3
asgiref==3.3.1
billiard==3.6.3.0
blis==0.7.3
CacheControl==0.12.6
catalogue==1.0.0
celery==5.0.3
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
cymem==2.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
joblib==0.17.0
kombu==5.0.2
lockfile==0.12.2
msgpack==0.6.2
murmurhash==1.0.4
numpy==1.19.4
packaging==20.3
pep517==0.8.2
plac==1.1.3
preshed==3.0.4
progress==1.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.4
rabbitmq==0.2.0
requests==2.22.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.4
six==1.14.0
spacy==2.3.4
sqlparse==0.4.1
srsly==1.0.4
thinc==7.4.3
threadpoolctl==2.1.0
tqdm==4.54.1
urllib3==1.25.8
vine==5.0.0
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
